Slide Generator/requirements.txt
streamlit
groq
python-pptx
python-dotenv
PyMuPDF
python-docx
Pillow